fix: util func refactor
